Issue275 pt. 1 MenuHelper simplification
