+++ app/pom.xml
@@ -237,12 +237,6 @@
             <version>${struts2.version}</version>
         </dependency>
 
-        <dependency>
-            <groupId>xml-security</groupId>
-            <artifactId>xmlsec</artifactId>
-            <version>1.3.0</version>
-        </dependency>
-
         <!-- Java EE deps -->
 
         <dependency>
