+++ app/src/main/java/org/apache/roller/weblogger/ui/core/menu/MenuTab.java
@@ -16,29 +16,58 @@
  * directory of this distribution.
  */
 
-package org.apache.roller.weblogger.ui.core.util.menu;
+package org.apache.roller.weblogger.ui.core.menu;
 
 import java.util.ArrayList;
 import java.util.List;
 
 
 /**
- * A parsed xml defined menu.
+ * Tab in a Menu.
  */
-public class ParsedMenu {
+public class MenuTab {
     
-    private List<ParsedTab> tabs = new ArrayList<ParsedTab>();
+    private String key = null;
+    private String action = null;
+    private boolean selected = false;
+    private List<MenuTabItem> items = new ArrayList<MenuTabItem>();
     
-    public void addTab(ParsedTab tab) {
-        this.tabs.add(tab);
+    
+    public void addItem(MenuTabItem item) {
+        this.items.add(item);
+    }
+    
+    
+    public String getKey() {
+        return key;
     }
 
-    public List<ParsedTab> getTabs() {
-        return tabs;
+    public void setKey(String key) {
+        this.key = key;
     }
 
-    public void setTabs(List<ParsedTab> tabs) {
-        this.tabs = tabs;
+    public String getAction() {
+        return action;
+    }
+
+    public void setAction(String url) {
+        this.action = url;
+    }
+
+    public boolean isSelected() {
+        return selected;
+    }
+
+    public void setSelected(boolean selected) {
+        this.selected = selected;
+    }
+    
+    public List<MenuTabItem> getItems() {
+        return items;
+    }
+
+    public void setItems(List<MenuTabItem> items) {
+        this.items = items;
     }
     
 }
