+++ app/src/main/java/org/apache/roller/weblogger/ui/core/menu/ParsedTabItem.java
@@ -16,12 +16,11 @@
  * directory of this distribution.
  */
 
-package org.apache.roller.weblogger.ui.core.util.menu;
+package org.apache.roller.weblogger.ui.core.menu;
 
 import org.apache.roller.weblogger.pojos.GlobalRole;
 import org.apache.roller.weblogger.pojos.WeblogRole;
 
-import java.util.List;
 import java.util.Set;
