+++ app/src/main/java/org/apache/roller/weblogger/ui/rendering/util/WeblogFeedRequest.java
@@ -32,7 +32,6 @@ import org.apache.roller.weblogger.business.WeblogManager;
 import org.apache.roller.weblogger.config.WebloggerConfig;
 import org.apache.roller.weblogger.business.WebloggerFactory;
 import org.apache.roller.weblogger.pojos.WeblogCategory;
-import org.apache.roller.weblogger.util.URLUtilities;
 import org.apache.roller.weblogger.util.Utilities;
 
 
@@ -125,7 +124,7 @@ public class WeblogFeedRequest extends WeblogRequest {
         if(request.getParameter("cat") != null) {
             // replacing plus sign below with its encoded equivalent (http://stackoverflow.com/a/6926987)
             this.weblogCategoryName =
-                    URLUtilities.decode(request.getParameter("cat").replace("+", "%2B"));
+                    Utilities.decode(request.getParameter("cat").replace("+", "%2B"));
         }
         
         if(request.getParameter("tags") != null) {
@@ -151,7 +150,7 @@ public class WeblogFeedRequest extends WeblogRequest {
         
         if(request.getParameter("q") != null &&
                 request.getParameter("q").trim().length() > 0) {
-            this.term = URLUtilities.decode(request.getParameter("q"));
+            this.term = Utilities.decode(request.getParameter("q"));
         }        
         
         if((this.tags != null && this.tags.size() > 0) && this.weblogCategoryName != null) {
