+++ app/src/main/java/org/apache/roller/weblogger/ui/rendering/util/WeblogPreviewRequest.java
@@ -29,7 +29,7 @@ import org.apache.roller.weblogger.business.themes.ThemeManager;
 import org.apache.roller.weblogger.business.WeblogEntryManager;
 import org.apache.roller.weblogger.pojos.Theme;
 import org.apache.roller.weblogger.pojos.WeblogEntry;
-import org.apache.roller.weblogger.util.URLUtilities;
+import org.apache.roller.weblogger.util.Utilities;
 
 
 /**
@@ -68,7 +68,7 @@ public class WeblogPreviewRequest extends WeblogPageRequest {
         
         // we may also have a specific entry to preview
         if(request.getParameter("previewEntry") != null) {
-            this.previewEntry = URLUtilities.decode(request.getParameter("previewEntry"));
+            this.previewEntry = Utilities.decode(request.getParameter("previewEntry"));
         }
 
         if(log.isDebugEnabled()) {
