+++ app/src/main/java/org/apache/roller/weblogger/ui/rendering/util/WeblogSearchRequest.java
@@ -24,9 +24,8 @@ import org.apache.commons.logging.LogFactory;
 import org.apache.roller.weblogger.WebloggerException;
 import org.apache.roller.weblogger.business.WeblogManager;
 import org.apache.roller.weblogger.business.WebloggerFactory;
-import org.apache.roller.weblogger.business.WeblogEntryManager;
 import org.apache.roller.weblogger.pojos.WeblogCategory;
-import org.apache.roller.weblogger.util.URLUtilities;
+import org.apache.roller.weblogger.util.Utilities;
 
 
 /**
@@ -99,7 +98,7 @@ public class WeblogSearchRequest extends WeblogRequest {
         if(request.getParameter("cat") != null &&
                 request.getParameter("cat").trim().length() > 0) {
             this.weblogCategoryName =
-                    URLUtilities.decode(request.getParameter("cat"));
+                    Utilities.decode(request.getParameter("cat"));
         }
     }
