+++ app/src/main/java/org/apache/roller/weblogger/util/Utilities.java
@@ -20,16 +20,10 @@
  */
 package org.apache.roller.weblogger.util;
 
-import java.io.BufferedInputStream;
-import java.io.BufferedOutputStream;
 import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.io.OutputStream;
 import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
 import java.net.URLEncoder;
@@ -37,8 +31,10 @@ import java.security.MessageDigest;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Locale;
+import java.util.Map;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -47,12 +43,10 @@ import javax.activation.MimetypesFileTypeMap;
 import javax.mail.internet.AddressException;
 import javax.mail.internet.InternetAddress;
 
-import org.apache.commons.codec.binary.Base64;
 import org.apache.commons.lang3.StringEscapeUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.apache.roller.weblogger.WebloggerCommon;
 
 /**
  * General purpose utilities, not for use in templates.
@@ -105,23 +99,6 @@ public class Utilities {
             "&lt;a href=.*?&gt;", Pattern.CASE_INSENSITIVE);
     private static final Pattern QUOTE_PATTERN = Pattern.compile("&quot;",
             Pattern.CASE_INSENSITIVE);
-
-    // ------------------------------------------------------------------------
-    /** Strip jsessionid off of a URL */
-    public static String stripJsessionId(String url) {
-        // Strip off jsessionid found in referer URL
-        int startPos = url.indexOf(";jsessionid=");
-        if (startPos != -1) {
-            int endPos = url.indexOf('?', startPos);
-            if (endPos == -1) {
-                url = url.substring(0, startPos);
-            } else {
-                url = url.substring(0, startPos)
-                        + url.substring(endPos, url.length());
-            }
-        }
-        return url;
-    }
 
     // ------------------------------------------------------------------------
     /**
@@ -215,22 +192,6 @@ public class Utilities {
 
     // ------------------------------------------------------------------------
     /**
-     * Run both removeHTML and escapeHTML on a string.
-     * 
-     * @param s
-     *            String to be run through removeHTML and escapeHTML.
-     * @return String with HTML removed and HTML special characters escaped.
-     */
-    public static String removeAndEscapeHTML(String s) {
-        if (s == null) {
-            return "";
-        } else {
-            return Utilities.escapeHTML(Utilities.removeHTML(s));
-        }
-    }
-
-    // ------------------------------------------------------------------------
-    /**
      * Autoformat.
      */
     public static String autoformat(String s) {
@@ -267,14 +228,6 @@ public class Utilities {
 
     // ------------------------------------------------------------------------
     /**
-     * Replaces occurences of non-alphanumeric characters with an underscore.
-     */
-    public static String replaceNonAlphanumeric(String str) {
-        return replaceNonAlphanumeric(str, '_');
-    }
-
-    // ------------------------------------------------------------------------
-    /**
      * Replaces occurrences of non-alphanumeric characters with a supplied char.
      */
     public static String replaceNonAlphanumeric(String str, char subst) {
@@ -291,22 +244,6 @@ public class Utilities {
     }
 
     // ------------------------------------------------------------------------
-    /**
-     * Remove occurences of non-alphanumeric characters.
-     */
-    public static String removeNonAlphanumeric(String str) {
-        StringBuilder ret = new StringBuilder(str.length());
-        char[] testChars = str.toCharArray();
-        for (int i = 0; i < testChars.length; i++) {
-            // MR: Allow periods in page links
-            if (Character.isLetterOrDigit(testChars[i]) || testChars[i] == '.') {
-                ret.append(testChars[i]);
-            }
-        }
-        return ret.toString();
-    }
-
-    // ------------------------------------------------------------------------
     /** Convert string array to string with delimeters. */
     public static String stringArrayToString(String[] stringArray, String delim) {
         StringBuilder bldr = new StringBuilder();
@@ -315,21 +252,6 @@ public class Utilities {
                 bldr.append(delim).append(stringArray[i]);
             } else {
                 bldr.append(stringArray[i]);
-            }
-        }
-        return bldr.toString();
-    }
-
-    // ------------------------------------------------------------------------
-    /** Convert string array to string with delimeters. */
-    public static String stringListToString(List<String> stringList,
-            String delim) {
-        StringBuilder bldr = new StringBuilder();
-        for (String s : stringList) {
-            if (bldr.length() > 0) {
-                bldr.append(delim).append(s);
-            } else {
-                bldr.append(s);
             }
         }
         return bldr.toString();
@@ -351,59 +273,6 @@ public class Utilities {
         return stringList;
     }
 
-    // --------------------------------------------------------------------------
-    /** Convert string to integer array. */
-    public static int[] stringToIntArray(String instr, String delim) {
-        String[] str = StringUtils.split(instr, delim);
-        int intArray[] = new int[str.length];
-        int i = 0;
-        for (String string : str) {
-            int nInt = Integer.parseInt(string);
-            intArray[i++] = nInt;
-        }
-        return intArray;
-    }
-
-    // -------------------------------------------------------------------
-    /** Convert integer array to a string. */
-    public static String intArrayToString(int[] intArray) {
-        StringBuilder bldr = new StringBuilder();
-        for (int s : intArray) {
-            if (bldr.length() > 0) {
-                bldr.append(",").append(Integer.toString(s));
-            } else {
-                bldr.append(Integer.toString(s));
-            }
-        }
-        return bldr.toString();
-    }
-
-    // ------------------------------------------------------------------------
-    public static void copyFile(File from, File to) throws IOException {
-        InputStream in;
-        OutputStream out;
-
-        try {
-            in = new FileInputStream(from);
-        } catch (IOException ex) {
-            throw new IOException("Utilities.copyFile: opening input stream '"
-                    + from.getPath() + "', " + ex.getMessage());
-        }
-
-        try {
-            out = new FileOutputStream(to);
-        } catch (Exception ex) {
-            try {
-                in.close();
-            } catch (IOException ex1) {
-            }
-            throw new IOException("Utilities.copyFile: opening output stream '"
-                    + to.getPath() + "', " + ex.getMessage());
-        }
-
-        copyInputToOutput(in, out, from.length());
-    }
-
     /**
      * Reads an inputstream into a string
      */
@@ -416,85 +285,6 @@ public class Utilities {
             sb.append(System.getProperty("line.separator"));
         }
         return sb.toString();
-    }
-
-    // ------------------------------------------------------------------------
-    /**
-     * Utility method to copy an input stream to an output stream. Wraps both
-     * streams in buffers. Ensures right numbers of bytes copied.
-     */
-    public static void copyInputToOutput(InputStream input,
-            OutputStream output, long byteCount) throws IOException {
-        int bytes;
-        long length;
-
-        BufferedInputStream in = new BufferedInputStream(input);
-        BufferedOutputStream out = new BufferedOutputStream(output);
-
-        byte[] buffer;
-        buffer = new byte[WebloggerCommon.EIGHT_KB_IN_BYTES];
-
-        for (length = byteCount; length > 0;) {
-            bytes = (int) (length > WebloggerCommon.EIGHT_KB_IN_BYTES ? WebloggerCommon.EIGHT_KB_IN_BYTES : length);
-
-            try {
-                bytes = in.read(buffer, 0, bytes);
-            } catch (IOException ex) {
-                try {
-                    in.close();
-                    out.close();
-                } catch (IOException ex1) {
-                }
-                throw new IOException("Reading input stream, "
-                        + ex.getMessage());
-            }
-
-            if (bytes < 0) {
-                break;
-            }
-
-            length -= bytes;
-
-            try {
-                out.write(buffer, 0, bytes);
-            } catch (IOException ex) {
-                try {
-                    in.close();
-                    out.close();
-                } catch (IOException ex1) {
-                }
-                throw new IOException("Writing output stream, "
-                        + ex.getMessage());
-            }
-        }
-
-        try {
-            in.close();
-            out.close();
-        } catch (IOException ex) {
-            throw new IOException("Closing file streams, " + ex.getMessage());
-        }
-    }
-
-    // ------------------------------------------------------------------------
-    public static void copyInputToOutput(InputStream input, OutputStream output)
-            throws IOException {
-        BufferedInputStream in = new BufferedInputStream(input);
-        BufferedOutputStream out = new BufferedOutputStream(output);
-        byte buffer[] = new byte[WebloggerCommon.EIGHT_KB_IN_BYTES];
-        for (int count = 0; count != -1;) {
-            count = in.read(buffer, 0, WebloggerCommon.EIGHT_KB_IN_BYTES);
-            if (count != -1) {
-                out.write(buffer, 0, count);
-            }
-        }
-
-        try {
-            in.close();
-            out.close();
-        } catch (IOException ex) {
-            throw new IOException("Closing file streams, " + ex.getMessage());
-        }
     }
 
     /**
@@ -543,36 +333,6 @@ public class Utilities {
         }
 
         return buf.toString();
-    }
-
-    /**
-     * Encode a string using Base64 encoding. Used when storing passwords as
-     * cookies.
-     * 
-     * This is weak encoding in that anyone can use the decodeString routine to
-     * reverse the encoding.
-     * 
-     * @param str
-     *            String to encode
-     * @return String
-     * @throws IOException
-     */
-    public static String encodeString(String str) throws IOException {
-        String encodedStr = new String(Base64.encodeBase64(str.getBytes()));
-        return (encodedStr.trim());
-    }
-
-    /**
-     * Decode a string using Base64 encoding.
-     * 
-     * @param str
-     *            String to decode
-     * @return decoded string
-     * @throws IOException
-     */
-    public static String decodeString(String str) throws IOException {
-        String value = new String(Base64.decodeBase64(str.getBytes()));
-        return (value);
     }
 
     /**
@@ -676,10 +436,6 @@ public class Utilities {
                 // get all the HTML from original str after loc
                 str3 = extractHTML(str.substring(loc));
 
-                // remove any tags which generate visible HTML
-                // This call is unecessary, all HTML has already been stripped
-                // str3 = removeVisibleHTMLTags(str3);
-
                 // append the appendToEnd String and
                 // add extracted HTML back onto truncated string
                 str = str2 + appendToEnd + str3;
@@ -727,103 +483,6 @@ public class Utilities {
     }
 
     /**
-     * @param str
-     * @return
-     */
-    private static String stripLineBreaks(String str) {
-        // TODO: use a string buffer, ignore case !
-        str = str.replaceAll("<br>", "");
-        str = str.replaceAll("<br/>", "");
-        str = str.replaceAll("<br />", "");
-        str = str.replaceAll("<p></p>", "");
-        str = str.replaceAll("<p/>", "");
-        str = str.replaceAll("<p />", "");
-        return str;
-    }
-
-    /**
-     * Need to get rid of any user-visible HTML tags once all text has been
-     * removed such as &lt;BR&gt;. This sounds like a better approach than
-     * removing all HTML tags and taking the chance to leave some tags
-     * un-closed.
-     * 
-     * WARNING: this method has serious performance problems a
-     * 
-     * @author Alexis Moussine-Pouchkine
-     *         (alexis.moussine-pouchkine@france.sun.com)
-     * @author Lance Lavandowska
-     * @param str
-     *            the String object to modify
-     * @return the new String object without the HTML "visible" tags
-     */
-    private static String removeVisibleHTMLTags(String str) {
-        str = stripLineBreaks(str);
-        StringBuilder result = new StringBuilder(str);
-        StringBuilder lcresult = new StringBuilder(str.toLowerCase());
-
-        // <img should take care of smileys, others to add?
-        String[] visibleTags = { "<img" };
-        int stringIndex;
-        for (int j = 0; j < visibleTags.length; j++) {
-            while ((stringIndex = lcresult.indexOf(visibleTags[j])) != -1) {
-                if (visibleTags[j].endsWith(">")) {
-                    result.delete(stringIndex,
-                            stringIndex + visibleTags[j].length());
-                    lcresult.delete(stringIndex,
-                            stringIndex + visibleTags[j].length());
-                } else {
-                    // need to delete everything up until next closing '>', for
-                    // <img for instance
-                    int endIndex = result.indexOf(">", stringIndex);
-                    if (endIndex > -1) {
-                        // only delete it if we find the end! If we don't the
-                        // HTML may be messed up, but we
-                        // can't safely delete anything.
-                        result.delete(stringIndex, endIndex + 1);
-                        lcresult.delete(stringIndex, endIndex + 1);
-                    }
-                }
-            }
-        }
-
-        // TODO: This code is buggy by nature. It doesn't deal with nesting of
-        // tags properly.
-        // remove certain elements with open & close tags, more available?
-        String[] openCloseTags = { "li", "a", "div", "h1", "h2", "h3", "h4" };
-        for (int j = 0; j < openCloseTags.length; j++) {
-            // could this be better done with a regular expression?
-            String closeTag = "</" + openCloseTags[j] + ">";
-            int lastStringIndex = 0;
-            while ((stringIndex = lcresult.indexOf("<" + openCloseTags[j],
-                    lastStringIndex)) > -1) {
-                lastStringIndex = stringIndex;
-                // Try to find the matching closing tag (ignores possible
-                // nesting!)
-                int endIndex = lcresult.indexOf(closeTag, stringIndex);
-                if (endIndex > -1) {
-                    // If we found it delete it.
-                    result.delete(stringIndex, endIndex + closeTag.length());
-                    lcresult.delete(stringIndex, endIndex + closeTag.length());
-                } else {
-                    // Try to see if it is a self-closed empty content tag, i.e.
-                    // closed with />.
-                    endIndex = lcresult.indexOf(">", stringIndex);
-                    int nextStart = lcresult.indexOf("<", stringIndex + 1);
-                    if (endIndex > stringIndex
-                            && lcresult.charAt(endIndex - 1) == '/'
-                            && (endIndex < nextStart || nextStart == -1)) {
-                        // Looks like it, so remove it.
-                        result.delete(stringIndex, endIndex + 1);
-                        lcresult.delete(stringIndex, endIndex + 1);
-                    }
-                }
-            }
-        }
-
-        return result.toString();
-    }
-
-    /**
      * Extract (keep) JUST the HTML from the String.
      * 
      * @param str
@@ -858,63 +517,6 @@ public class Utilities {
             }
         }
         return ret.toString();
-    }
-
-    /**
-     * URL encoding.
-     * 
-     * @param s
-     *            a string to be URL-encoded
-     * @return URL encoding of s using character encoding UTF-8; null if s is
-     *         null.
-     */
-    public static final String encode(String s) {
-        try {
-            if (s != null) {
-                return URLEncoder.encode(s, "UTF-8");
-            } else {
-                return s;
-            }
-        } catch (UnsupportedEncodingException e) {
-            // Java Spec requires UTF-8 be in all Java environments, so this
-            // should not happen
-            return s;
-        }
-    }
-
-    /**
-     * URL decoding.
-     * 
-     * @param s
-     *            a URL-encoded string to be URL-decoded
-     * @return URL decoded value of s using character encoding UTF-8; null if s
-     *         is null.
-     */
-    public static final String decode(String s) {
-        try {
-            if (s != null) {
-                return URLDecoder.decode(s, "UTF-8");
-            } else {
-                return s;
-            }
-        } catch (UnsupportedEncodingException e) {
-            // Java Spec requires UTF-8 be in all Java environments, so this
-            // should not happen
-            return s;
-        }
-    }
-
-    /**
-     * @param string
-     * @return
-     */
-    public static int stringToInt(String string) {
-        try {
-            return Integer.valueOf(string);
-        } catch (NumberFormatException e) {
-            mLogger.debug("Invalid Integer:" + string);
-        }
-        return 0;
     }
 
     /**
@@ -1116,4 +718,122 @@ public class Utilities {
                 && StringUtils.isNotEmpty(tokens[1]);
     }
 
+    /**
+     * Compose a map of key=value params into a query string.
+     */
+    public static String getQueryString(Map<String, String> params) {
+
+        if(params == null) {
+            return null;
+        }
+
+        StringBuilder queryString = new StringBuilder();
+        for (Map.Entry<String, String> entry : params.entrySet()) {
+
+            if (queryString.length() == 0) {
+                queryString.append("?");
+            } else {
+                queryString.append("&");
+            }
+
+            queryString.append(entry.getKey());
+            queryString.append("=");
+            queryString.append(entry.getValue());
+        }
+
+        return queryString.toString();
+    }
+
+    /**
+     * URL encode a string using UTF-8.
+     */
+    public static String encode(String str) {
+        String encodedStr = str;
+        try {
+            encodedStr = URLEncoder.encode(str, "UTF-8");
+        } catch (UnsupportedEncodingException ex) {
+            // ignored
+        }
+        return encodedStr;
+    }
+
+    /**
+     * URL decode a string using UTF-8.
+     */
+    public static String decode(String str) {
+        String decodedStr = str;
+        try {
+            decodedStr = URLDecoder.decode(str, "UTF-8");
+        } catch (UnsupportedEncodingException ex) {
+            // ignored
+        }
+        return decodedStr;
+    }
+
+    public static String getEncodedTagsString(List tags) {
+        StringBuilder tagsString = new StringBuilder();
+        if(tags != null && tags.size() > 0) {
+            String tag;
+            Iterator tagsIT = tags.iterator();
+
+            // do first tag
+            tag = (String) tagsIT.next();
+            tagsString.append(encode(tag));
+
+            // do rest of tags, joining them with a '+'
+            while(tagsIT.hasNext()) {
+                tag = (String) tagsIT.next();
+                tagsString.append("+");
+                tagsString.append(encode(tag));
+            }
+        }
+        return tagsString.toString();
+    }
+
+    /**
+     * URL encode a path string using UTF-8. The path separator '/' will not be encoded
+     */
+    public static String encodePath(String path) {
+        int i = path.indexOf('/');
+        StringBuilder sb = new StringBuilder();
+        while (i != -1) {
+            sb.append(encode(path.substring(0, i))).append('/');
+            path = path.substring(i + 1);
+            i = path.indexOf('/');
+        }
+        sb.append(encode(path));
+        return sb.toString();
+    }
+
+    public static Locale toLocale(String locale) {
+        if (locale != null) {
+            String[] localeStr = StringUtils.split(locale,"_");
+            if (localeStr.length == 1) {
+                if (localeStr[0] == null) {
+                    localeStr[0] = "";
+                }
+                return new Locale(localeStr[0]);
+            } else if (localeStr.length == 2) {
+                if (localeStr[0] == null) {
+                    localeStr[0] = "";
+                }
+                if (localeStr[1] == null) {
+                    localeStr[1] = "";
+                }
+                return new Locale(localeStr[0], localeStr[1]);
+            } else if (localeStr.length == 3) {
+                if (localeStr[0] == null) {
+                    localeStr[0] = "";
+                }
+                if (localeStr[1] == null) {
+                    localeStr[1] = "";
+                }
+                if (localeStr[2] == null) {
+                    localeStr[2] = "";
+                }
+                return new Locale(localeStr[0], localeStr[1], localeStr[2]);
+            }
+        }
+        return Locale.getDefault();
+    }
 }
