+++ app/src/main/resources/org/apache/roller/weblogger/ui/struts2/admin/admin-menu.xml
@@ -23,9 +23,9 @@ See comments in editor-menu.xml for info about attributes.
                    subactions="commonPingTargetAdd,commonPingTargetEdit" />
 
         <menu-item action="planets"
-                   name="tabbedmenu.admin.planets" enabledProperty="planet.aggregator.enabled"
+                   name="tabbedmenu.admin.planets"
                    subactions="planetSubscriptions"
-                   />
+                   enabledProperty="planet.aggregator.enabled" />
 
         <menu-item action="cacheInfo"
                    name="tabbedmenu.admin.cacheInfo" />
