+++ app/src/test/java/org/apache/roller/weblogger/util/ExpiringLRUCacheImplTest.java
@@ -14,19 +14,20 @@
  * limitations under the License.  For additional information regarding
  * copyright in this work, please see the NOTICE file in the top level
  * directory of this distribution.
+ *
+ * Source file modified from the original ASF source; all changes made
+ * are also under Apache License.
  */
-
 package org.apache.roller.weblogger.util;
 
 import junit.framework.Test;
 import junit.framework.TestCase;
 import junit.framework.TestSuite;
+import org.apache.roller.weblogger.util.cache.Cache;
+import org.apache.roller.weblogger.util.cache.ExpiringLRUCacheImpl;
 
 
-/**
- * Test LRUCache2.
- */
-public class LRUCache2Test extends TestCase {
+public class ExpiringLRUCacheImplTest extends TestCase {
     
     /**
      * @see junit.framework.TestCase#setUp()
@@ -36,31 +37,10 @@ public class LRUCache2Test extends TestCase {
         super.setUp();
     }
     
-    public void testTimeout() {
-        // Create cache with 100 item limit and 15 second timeout
-        TestEnvironment env = new TestEnvironment();
-        LRUCache2 cache = new LRUCache2(env, 100, 15000);
-        
-        env.time = 1000;
-        cache.put("key1", "string1");
-        cache.put("key2", "string2");
-        cache.put("key3", "string3");
-        assertNotNull(cache.get("key1"));
-        assertNotNull(cache.get("key2"));
-        assertNotNull(cache.get("key3"));
-        
-        env.time = 16000;
-        assertNull(cache.get("key1"));
-        assertNull(cache.get("key2"));
-        assertNull(cache.get("key3"));
-    }
-    
     public void testLRU() {
         // Create cache with 3 item limit and 15 second timeout
-        TestEnvironment env = new TestEnvironment();
-        LRUCache2 cache = new LRUCache2(env, 3, 15000);
+        Cache cache = new ExpiringLRUCacheImpl("test", 3, 15000);
         
-        env.time = 1000;
         cache.put("key1", "string1");
         cache.put("key2", "string2");
         cache.put("key3", "string3");
@@ -79,12 +59,10 @@ public class LRUCache2Test extends TestCase {
         assertNull(cache.get("key3"));
     }
     
-    public void testPurge() {
+    public void testRemove() {
         // Create cache with 100 item limit and 15 second timeout
-        TestEnvironment env = new TestEnvironment();
-        LRUCache2 cache = new LRUCache2(env, 100, 15000);
-        
-        env.time = 1000;
+        Cache cache = new ExpiringLRUCacheImpl("test", 100, 15000);
+
         cache.put("key1", "string1");
         cache.put("key2", "string2");
         cache.put("key3", "string3");
@@ -92,11 +70,14 @@ public class LRUCache2Test extends TestCase {
         assertNotNull(cache.get("key2"));
         assertNotNull(cache.get("key3"));
         
-        cache.purge(new String[] {"key1", "key2"});
-        assertEquals(1, cache.size());
-        
-        cache.purge();
-        assertEquals(0, cache.size());
+        cache.remove("key1");
+        cache.remove("key2");
+        assertNull(cache.get("key1"));
+        assertNull(cache.get("key2"));
+        assertNotNull(cache.get("key3"));
+
+        cache.clear();
+        assertNull(cache.get("key3"));
     }
     
     /**
@@ -107,15 +88,8 @@ public class LRUCache2Test extends TestCase {
         super.tearDown();
     }
     
-    public static class TestEnvironment implements LRUCache2.Environment {
-        public long time = 0;
-        public long getCurrentTimeInMillis() {
-            return time;
-        }
-    }
-    
     public static Test suite() {
-        return new TestSuite(LRUCache2Test.class);
+        return new TestSuite(ExpiringLRUCacheImplTest.class);
     }
     
 }
